Add datastructure for I-compartment
Function create-I-channels creates a map. Each enttry is a map holding
number of times the channel has been touched, number of touches that are
required, number of infected individuals and a channel to-cohort. When
the channel has been touched required number of times, the number of
infected individuals is put on the channel to-cohort.