Add test for progression and fix progression
There is now a test that checks if the numbers in the compartments over
the course of the progression of a single cohort add up to the numbers
they should. Adding all compartments should always give the same number.
Also, the number of susceptibles within a cohort should always stay the
same, since after the start of a cohort, where new infections happen,
during the progression there is only a flow from the I to the R
compartment.

The test now passes. Fixed errors in the updating of the compartments.
Specifically, removed at t = removed at t-1 + new removed at t.

Also removed unnecessary function in compartments-namespace. All the
update-X functions are gone. Instead, call assoc or assoc-in directly
where needed.