Change start-cohort-function
The start of a cohort is now a two-phased process. First the already
infected generate new, primary cases. Then these individuals infect new,
secondary cases. The primary and secondary cases together make up the
new weekly cases.

The number of new cases for an individual is now a poisson distributed
random variable. The parameters for the already-infected and the newly-infected should not be the same. The idea is, that newly-infected people infect more people that the already-infected.

All test in the cohort and core namespaces so far pass.