Functions to add values to compartments
To be used in the progression of a cohort. The thing to realize is that
the number of people that drop out during the progression and move to
the R-compartment, as well as those that remain infected and stay in the
R-compartment is initially computed as the sum of primary and
secondary compartments. After that though, the number of dropputs is
only available inside the function that simulates the cohort
progression. This function, however, does access the compartments at
every timestep to add removed cases to R and remaining cases to I (for
the next timestep).