Modularize cohort progressions
The whole process of formation and progeression of a cohort is split
into smaller fuunctions. There are different levels of abstraction. On
the highest level, a lifetime-fn is passed as an argument to to query.
The lifetime-fn semantically consists of a formation-fn, which marks the
start of a cohort, and a progression-fn, which takes care of the
progression. So far, though, the splitting into formation-fn and
progression-fn is not in any way forced or checked programmaticaly, but
a conceptual idea.

The formation-fn, which is called start-poisson-poisson, has two phases
represented by two functions called primary-poisson and
secondary-poisson. The progression-fn, which is called progress, is not
split further. The implementation of the lifetim-fn is called
form-and-prog.