Plot aggregates
Mean, median and mode are now plotted nicely on top of new infections. A
new plot with each of these statistics is drawn. Also, weeky
distributions are now plotted with the scales=free argument and without.
This way it becomes possibly to compare the central tendencies as well
as how narrow or wide the distributions are.