Major Refactoring
Add new namespace "framework" that (so far) only has the function
season-fn. This function does not take care of initializing the
compartments anymore. This part is moved to the function
init-compartments in the "two-stage-poisson" namespace.

Generally, everything that is related to the model is now in the
"two-stage-poisson" namespace. Everything that is more generic and is
independent of the model should go to the "framework" namespace. Because
of this, when calling doquery, no function for the lifetime of a cohort
is passed as an argument anymore. This lifetime function is considered
part of the model, lives in the namespace and is therefore just used
inside of the defquery (which is itself related to the model and lives
in the model namespace. Args are still passed as an argument.

When starting the simulation from the "core" namespace, one now only has
to call doquery on this query. This way, a different model can easily be
developed in a different namespace and be called by specifying its query
(and optional args).